feat(reserva): add overbooking flag, centralize Devolucion creation and use nullable comprobante for pagos

+++ database/migrations/2025_10_20_000002_make_idreserva_nullable_in_pagos_table.php
@@ -8,14 +8,14 @@ return new class extends Migration {
     public function up(): void
     {
         Schema::table('pagos', function (Blueprint $table) {
-            $table->integer('idReserva')->nullable()->change();
+            $table->unsignedBigInteger('idReserva')->nullable()->change();
         });
     }
 
     public function down(): void
     {
         Schema::table('pagos', function (Blueprint $table) {
-            $table->integer('idReserva')->nullable(false)->change();
+            $table->unsignedBigInteger('idReserva')->nullable(false)->change();
         });
     }
 };
